Simplify injection logic to GenericStack
